register_amethyst_type option to not wrap structure in dict when encoding objects to json
If an Attr() can inflate an object from plain structure (list, dict, string), then it is
not necessary to wrap its encoded form in a dict of { __class__: struct }. In fact, we may
not want to do this in case the JSON structure is loaded before we get it (in which case
our object-hook is never used since we plug into JSON inflation directly -- may want to
re-evaluate this at some point). This patch adds an option to register_amethyst_type to
avoid wrapping the encoded structure to rely on the usual Attr() `convert` callback to
inflate the structure.