flag, minimize, and codify core API
Object() has a number of methods which are convenient, but not required for
core functionality. In order to reduce namespace pollution, all core
required methods have been prefixed with "amethyst_". Other methods have
been explicitly marked as safe for subclasses to overwrite with possibly
unrelated behavior. Tests have been added to ensure this remains so. The
only ones remaining are the camel-case JSON methods which are overridable
in whole, but not recommended.